helper funtion

+++ fsoares/file_utils.c
@@ -6,10 +6,10 @@
 
 int next_read_error = 0;
 
-ssize_t read(int fildes, void *buf, size_t nbyte)
+ssize_t read(int fildes, void *buf, int nbyte)
 {
-	ssize_t(*libc_read)(int, void *, size_t) =
-		(ssize_t (*)(int, void *, size_t))dlsym(RTLD_NEXT, "read");
+	ssize_t(*libc_read)(int, void *, int) =
+		(ssize_t (*)(int, void *, int))dlsym(RTLD_NEXT, "read");
 	if (nbyte != BUFFER_SIZE && nbyte != 0) {
 		fprintf(errors_file, "Please use the read size provided by the compile variable BUFFER_SIZE: %i\n", BUFFER_SIZE);
 		printf(RED "Use the BUFFER_SIZE (%i) on the read calls, instead of what you wanted to read (%zi).\n" NC, BUFFER_SIZE, nbyte);
